test: convert runner to node:test, hardening
- fix(add_line_processor): add MAX_LINE_LENGTH bounds
- fix(get_public_ip): assure STUN response is a valid IP
- fix(get_public_ip): add null guard to config.get
- test: convert runner to node:test
