+++ index.js
@@ -95,33 +95,17 @@ exports.is_ip_in_str = function (ip, str) {
   return false
 }
 
-const re_ipv4 = {
-  loopback: /^127\./,
-  link_local: /^169\.254\./,
+const {
+  is_private_ipv4,
+  is_local_ipv4,
+  is_local_ipv6,
+  is_private_ip,
+} = require('./lib/ip')
 
-  private10: /^10\./, // 10/8
-  private192: /^192\.168\./, // 192.168/16
-  // 172.16/16 .. 172.31/16
-  private172: /^172\.(1[6-9]|2[0-9]|3[01])\./, // 172.16/12
-
-  // RFC 5735
-  testnet1: /^192\.0\.2\./, // 192.0.2.0/24
-  testnet2: /^198\.51\.100\./, // 198.51.100.0/24
-  testnet3: /^203\.0\.113\./, // 203.0.113.0/24
-}
-
-exports.is_private_ipv4 = function (ip) {
-  // RFC 1918, reserved as "private" IP space
-  if (re_ipv4.private10.test(ip)) return true
-  if (re_ipv4.private192.test(ip)) return true
-  if (re_ipv4.private172.test(ip)) return true
-
-  if (re_ipv4.testnet1.test(ip)) return true
-  if (re_ipv4.testnet2.test(ip)) return true
-  if (re_ipv4.testnet3.test(ip)) return true
-
-  return false
-}
+exports.is_private_ipv4 = is_private_ipv4
+exports.is_local_ipv4 = is_local_ipv4
+exports.is_local_ipv6 = is_local_ipv6
+exports.is_private_ip = is_private_ip
 
 exports.on_local_interface = function (ip) {
   if (locallyBoundIPs.length === 0) {
@@ -172,52 +156,10 @@ exports.is_local_host = async function (dst_host) {
 exports.is_local_ip = function (ip) {
   if (exports.on_local_interface(ip)) return true
 
-  if (net.isIPv4(ip)) return exports.is_local_ipv4(ip)
-  if (net.isIPv6(ip)) return exports.is_local_ipv6(ip)
+  if (net.isIPv4(ip)) return is_local_ipv4(ip)
+  if (net.isIPv6(ip)) return is_local_ipv6(ip)
 
   // console.error(`invalid IP address: ${ip}`);
-  return false
-}
-
-exports.is_local_ipv4 = function (ip) {
-  if ('0.0.0.0' === ip) return true // RFC 5735
-
-  // 127/8 (loopback)   # RFC 1122
-  if (re_ipv4.loopback.test(ip)) return true
-
-  // link local: 169.254/16 RFC 3927
-  if (re_ipv4.link_local.test(ip)) return true
-
-  return false
-}
-
-const re_ipv6 = {
-  loopback: /^(0{1,4}:){7}0{0,3}1$/,
-  link_local: /^fe80::/i,
-  unique_local: /^f(c|d)[a-f0-9]{2}:/i,
-}
-
-exports.is_local_ipv6 = function (ip) {
-  if (ip === '::') return true // RFC 5735
-  if (ip === '::1') return true // RFC 4291
-
-  // 2 more IPv6 notations for ::1
-  // 0:0:0:0:0:0:0:1 or 0000:0000:0000:0000:0000:0000:0000:0001
-  if (re_ipv6.loopback.test(ip)) return true
-
-  // link local: fe80::/10, RFC 4862
-  if (re_ipv6.link_local.test(ip)) return true
-
-  // unique local (fc00::/7)   -> fc00: - fd00:
-  if (re_ipv6.unique_local.test(ip)) return true
-
-  return false
-}
-
-exports.is_private_ip = function (ip) {
-  if (net.isIPv4(ip))
-    return exports.is_local_ipv4(ip) || exports.is_private_ipv4(ip)
-  if (net.isIPv6(ip)) return exports.is_local_ipv6(ip)
   return false
 }
 
@@ -353,12 +295,21 @@ exports.get_public_ip_async = require('./lib/get_public_ip').get_public_ip_async
 
 exports.HarakaMx = require('./lib/HarakaMx')
 
+const MAX_LINE_LENGTH = 4 * 1024 * 1024 // 4 MB; defence against DoS via lines without newlines
+
 exports.add_line_processor = (socket) => {
-  const line_regexp = /^([^\n]*\n)/ // utils.line_regexp
+  const line_regexp = /^([^\n]*\n)/
   let current_data = ''
 
   socket.on('data', (data) => {
     current_data += data
+
+    if (current_data.length > MAX_LINE_LENGTH) {
+      socket.emit('error', new Error(`Line length exceeded ${MAX_LINE_LENGTH} bytes`))
+      current_data = ''
+      return
+    }
+
     let results
     while ((results = line_regexp.exec(current_data))) {
       const this_line = results[1]
