+++ lib/get_mx.js
@@ -88,40 +88,25 @@ function normalizeDomain(raw_domain) {
 
   if (/@/.test(domain)) {
     domain = domain.split('@').pop()
-    // console.log(`\treduced ${raw_domain} to ${domain}.`)
   }
 
   if (/^xn--/.test(domain)) {
     // is punycode IDN with ACE, ASCII Compatible Encoding
   } else if (domain !== punycode.toASCII(domain)) {
     domain = punycode.toASCII(domain)
-    console.log(`\tACE encoded '${raw_domain}' to '${domain}'`)
+    console.log(`\tACE encoded '${raw_domain.replace(/[\r\n]/g, '')}' to '${domain}'`)
   }
 
   return domain
 }
 
 function fatal_mx_err(err) {
-  // Possible DNS errors
-  // NODATA
-  // FORMERR
-  // BADRESP
-  // NOTFOUND
-  // BADNAME
-  // TIMEOUT
-  // CONNREFUSED
-  // NOMEM
-  // DESTRUCTION
-  // NOTIMP
-  // EREFUSED
-  // SERVFAIL
-
   switch (err.code) {
     case 'ENODATA':
     case 'ENOTFOUND':
       // likely a hostname with no MX record, drop through
       return false
     default:
-      return err
+      return true
   }
 }
