+++ lib/get_public_ip.js
@@ -1,12 +1,15 @@
 'use strict'
 
+const net = require('node:net')
+const { is_private_ip } = require('./ip')
+
 exports.config = require('haraka-config')
 
 exports.get_public_ip_async = async function () {
   if (this.public_ip !== undefined) return this.public_ip // cache
 
   // manual config override, for the cases where we can't figure it out
-  const smtpIni = exports.config.get('smtp.ini').main
+  const smtpIni = exports.config.get('smtp.ini')?.main ?? {}
   if (smtpIni.public_ip) {
     this.public_ip = smtpIni.public_ip
     return this.public_ip
@@ -33,7 +36,12 @@ exports.get_public_ip_async = async function () {
   const res = await this.stun.request(get_stun_server(), {
     maxTimeout: (timeout - 1) * 1000,
   })
-  this.public_ip = res.getXorAddress().address
+  const ip = res.getXorAddress().address
+  if (net.isIP(ip) && !is_private_ip(ip)) {
+    this.public_ip = ip
+  } else {
+    console.error(`STUN returned unusable IP: ${ip}`)
+  }
   clearTimeout(timer)
   return this.public_ip
 }
@@ -44,7 +52,7 @@ exports.get_public_ip = async function (cb) {
   if (this.public_ip !== undefined) return cb(null, this.public_ip) // cache
 
   // manual config override, for the cases where we can't figure it out
-  const smtpIni = exports.config.get('smtp.ini').main
+  const smtpIni = exports.config.get('smtp.ini')?.main ?? {}
   if (smtpIni.public_ip) {
     this.public_ip = smtpIni.public_ip
     return cb(null, this.public_ip)
@@ -75,7 +83,12 @@ exports.get_public_ip = async function (cb) {
       if (timer) clearTimeout(timer)
       if (error) return cb(error)
 
-      this.public_ip = res.getXorAddress().address
+      const ip = res.getXorAddress().address
+      if (net.isIP(ip) && !is_private_ip(ip)) {
+        this.public_ip = ip
+      } else {
+        console.error(`STUN returned unusable IP: ${ip}`)
+      }
       cb(null, this.public_ip)
     },
   )
